+++ src/modules/health/health.controller.ts
@@ -10,20 +10,15 @@ import { Role } from "../users/models/role.enum";
 export class HealthController {
   constructor(private healthService: HealthService) {}
 
-  /**
-   * Returns a lifesign.
-   *
-   * @returns
-   */
   @Get()
   @ApiOkResponse({ type: () => Health })
   @ApiOperation({
     summary:
       "returns a lifesign, if an admin calls this api additional server infos are returned.",
-    operationId: "healthcheck",
+    operationId: "getHealth",
   })
   @Public()
-  async healthcheck(): Promise<Health> {
+  async getHealth(): Promise<Health> {
     return this.healthService.get();
   }
 
@@ -31,11 +26,12 @@ export class HealthController {
   @ApiOkResponse({ type: () => Health })
   @ApiOperation({
     summary:
-      "returns a lifesign and additional server metrics for administrators",
-    operationId: "healthcheck",
+      "returns a lifesign and additional server metrics for administrators. **DEPRECATED. MOVED TO /admin/health**",
+    operationId: "getHealthAdmin",
+    deprecated: true,
   })
   @MinimumRole(Role.ADMIN)
-  async extensiveHealthcheck(): Promise<Health> {
+  async getHealthAdmin(): Promise<Health> {
     return this.healthService.getExtensive();
   }
 }