+++ src/modules/users/models/activity-state.enum.ts
@@ -1,0 +1,5 @@
+export enum ActivityState {
+  ONLINE = "ONLINE",
+  BUSY = "BUSY",
+  PLAYING = "PLAYING",
+}